Put data types into a public module
